backend/requirements.txt
fastapi>=0.110,<0.115
uvicorn[standard]>=0.24,<0.32
SQLAlchemy>=2.0,<2.1
pydantic>=2.5,<3.0
pydantic-settings>=2.0,<3.0
python-jose>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6,<0.0.10
pymysql>=1.0.0
cryptography>=3.4.8
